Refactor sql actions
