+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/actions/data/sql/actions/Init.java
@@ -15,18 +15,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.nativeimpl.actions.data.sql.client;
+package org.ballerinalang.nativeimpl.actions.data.sql.actions;
 
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BConnector;
-import org.ballerinalang.model.values.BEnumerator;
-import org.ballerinalang.model.values.BMap;
-import org.ballerinalang.model.values.BString;
-import org.ballerinalang.model.values.BStruct;
+//import org.ballerinalang.model.values.BConnector;
+//import org.ballerinalang.model.values.BEnumerator;
+//import org.ballerinalang.model.values.BMap;
+//import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.nativeimpl.actions.data.sql.Constants;
-import org.ballerinalang.nativeimpl.actions.data.sql.SQLDatasource;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaAction;
 
@@ -48,21 +46,21 @@ public class Init extends AbstractSQLAction {
 
     @Override
     public void execute(Context context) {
-        BConnector bConnector = (BConnector) context.getRefArgument(0);
-        BEnumerator db = (BEnumerator) bConnector.getRefField(0);
-        BStruct optionStruct = (BStruct) bConnector.getRefField(1);
-        BMap sharedMap = (BMap) bConnector.getRefField(2);
-        String dbType = db.getName();
-        String hostOrPath = bConnector.getStringField(0);
-        String dbName = bConnector.getStringField(1);
-        String username = bConnector.getStringField(2);
-        String password = bConnector.getStringField(3);
-        int port = (int) bConnector.getIntField(0);
-        if (sharedMap.get(new BString(Constants.DATASOURCE_KEY)) == null) {
-            SQLDatasource datasource = new SQLDatasource();
-            datasource.init(optionStruct, dbType, hostOrPath, port, username, password, dbName);
-            sharedMap.put(new BString(Constants.DATASOURCE_KEY), datasource);
-        }
+//        BConnector bConnector = (BConnector) context.getRefArgument(0);
+//        BEnumerator db = (BEnumerator) bConnector.getRefField(0);
+//        BStruct optionStruct = (BStruct) bConnector.getRefField(1);
+//        BMap sharedMap = (BMap) bConnector.getRefField(2);
+//        String dbType = db.getName();
+//        String hostOrPath = bConnector.getStringField(0);
+//        String dbName = bConnector.getStringField(1);
+//        String username = bConnector.getStringField(2);
+//        String password = bConnector.getStringField(3);
+//        int port = (int) bConnector.getIntField(0);
+//        if (sharedMap.get(new BString(Constants.DATASOURCE_KEY)) == null) {
+//            SQLDatasource datasource = new SQLDatasource();
+//            datasource.init(optionStruct, dbType, hostOrPath, port, username, password, dbName);
+//            sharedMap.put(new BString(Constants.DATASOURCE_KEY), datasource);
+//        }
 
         context.setReturnValues();
     }
